shared/services/cms/requirements.txt
Django>=4.2,<5.0
wagtail>=5.2
psycopg2-binary>=2.9.0
redis>=4.5.0
requests>=2.31.0
python-decouple>=3.8
celery>=5.3.0
gunicorn>=21.2.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
djangorestframework>=3.14.0
PyJWT>=2.8.0
cryptography>=41.0.0
wagtail-modeladmin>=2.0.0
whitenoise>=6.6.0
django-extensions>=3.2.0
django-storages>=1.14.0
boto3>=1.34.0
hvac>=1.2.1
django-redis>=5.4.0